+++ src/v1/SupportingTypes/DispatchMap/AnyDispatchMapKey.ts
@@ -1,5 +1,5 @@
 //
-// Copyright (c) 2021-present Ganbaro Digital Ltd
+// Copyright (c) 2022-present Ganbaro Digital Ltd
 // All rights reserved.
 //
 // Redistribution and use in source and binary forms, with or without
@@ -32,8 +32,4 @@
 // POSSIBILITY OF SUCH DAMAGE.
 //
 
-export * from "./AnyDispatchMapKey";
-export * from "./AnyDispatchMap";
-export * from "./DispatchMap";
-export * from "./DispatchMapReturnTypes";
-export * from "./searchDispatchMap";
+export type AnyDispatchMapKey = number | string | symbol;