+++ CHANGELOG.md
@@ -21,6 +21,10 @@ For each release, changes are grouped under these headings:
 
 The following changes have been completed, and will be included in the next tagged release.
 
+### New
+
+* Added `Nullable` utility type.
+
 ## v0.3.5
 
 Released Saturday, 18th September 2021.