+++ CHANGELOG.md
@@ -24,6 +24,7 @@ The following changes have been completed, and will be included in the next tagg
 ### New
 
 * Added `Nullable` utility type.
+* Added `NonNullable` utility type.
 
 ## v0.3.5